Sistema/API/app/requirements.txt
alembic==1.6.2
click==8.0.0
colorama==0.4.4
DateTime==4.3
Flask==2.0.0
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==3.0.0
flask-redis==0.4.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
PyJWT==2.1.0
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
redis==3.5.3
six==1.16.0
SQLAlchemy==1.4.15
SQLAlchemy-Utils==0.37.3
uuid==1.30
Werkzeug==2.0.1
zope.interface==5.4.0
